template-creation-0-1.ps1
first popup you need to define the folder name where the templates are copied. After this you can browse to the folder where you want to create the named folder within templates. Do not create an folder on your own and choose this one. This script will do it for you.